♻️ フェーズ3: 削除処理などの共通化
